+++ server/Base/ServerMgr.h
@@ -37,8 +37,8 @@ public:
 	void ProcessNotRegister(serverinfo *info);
 
 	// 发送消息
-	void SendMsgToServer(serverinfo *con, google::protobuf::Message &pMsg, int maintype, int subtype, int64 nClientID = 0);
-	void SendMsgToServer(serverinfo *con, Msg &pMsg, int64 nClientID = 0);
+	void SendMsgToServer(serverinfo *con, google::protobuf::Message &pMsg, int maintype, int subtype, int32 nClientID = 0);
+	void SendMsgToServer(serverinfo *con, Msg &pMsg, int32 nClientID = 0);
 
 	void SendMsg(serverinfo *info, google::protobuf::Message &pMsg, int maintype, int subtype, void *adddata = nullptr, size_t addsize = 0);
 	void SendMsg(serverinfo *info, Msg &pMsg, void *adddata = nullptr, size_t addsize = 0);
@@ -46,8 +46,8 @@ public:
 	// 处理已注册服务器消息
 	virtual void ProcessMsg(serverinfo *info) = 0;
 	// 根据类型和SvrID发送消息给服务器
-	virtual void SendMsgToServer(google::protobuf::Message &pMsg, int maintype, int subtype, int nType, int64 nClientID = 0, int nServerID = 0, bool bBroad = false) = 0;
-	virtual void SendMsgToServer(Msg &pMsg, int nType, int64 nClientID = 0, int nServerID = 0, bool bBroad = false) = 0;
+	virtual void SendMsgToServer(google::protobuf::Message &pMsg, int maintype, int subtype, int nType, int32 nClientID = 0, int nServerID = 0, bool bBroad = false) = 0;
+	virtual void SendMsgToServer(Msg &pMsg, int nType, int32 nClientID = 0, int nServerID = 0, bool bBroad = false) = 0;
 private:
 	// 停止监听
 	void StopListen();
